refactor: stopover 대신 ShuttleRouteHub 으로 네이밍
